feat(recording): StartRecordingDialogContent web and native
