+++ react/features/recording/components/Recording/web/StartRecordingDialog.js
@@ -11,7 +11,9 @@ import { RECORDING_TYPES } from '../../../constants';
 import AbstractStartRecordingDialog, {
     mapStateToProps as abstractMapStateToProps
 } from '../AbstractStartRecordingDialog';
-import StartRecordingDialogContent from '../StartRecordingDialogContent';
+
+import StartRecordingDialogContent from './StartRecordingDialogContent';
+
 
 /**
  * React Component for getting confirmation to start a file recording session in
